refactor(Project): rename to `fold-wrapper`, remove Baseline models
