+++ LICENSE
@@ -28,7 +28,7 @@ Without limiting other conditions in the License, the grant of rights under the 
 
 For purposes of the foregoing, “Sell” means practicing any or all of the rights granted to you under the License to provide to third parties, for a fee or other consideration (including without limitation fees for hosting or consulting/ support services related to the Software), a product or service whose value derives, entirely or substantially, from the functionality of the Software. Any license notice or attribution required by the License must also include this Commons Clause License Condition notice.
 
-Software: fold-models (https://www.github.com/dream-faster/fold-models)
+Software: fold-wrapper (https://www.github.com/dream-faster/fold-wrapper)
 
 License: MIT License with Commons Clause
