+++ pyproject.toml
@@ -3,7 +3,7 @@ build-backend = "hatchling.build"
 requires = ["hatchling"]
 
 [project]
-name = "fold-models"
+name = "fold-wrapper"
 version = "0.0.1"
 authors = [
   { name="Mark Aron Szulyovszky", email="mark@dreamfaster.ai" },
@@ -32,9 +32,9 @@ readme = "README.md"
 requires-python = ">=3.7"
 
 [project.urls]
-Documentation = "https://dream-faster.github.io/fold-models"
-Issues = "https://github.com/dream-faster/fold-models/issues"
-Source = "https://github.com/dream-faster/fold-models"
+Documentation = "https://dream-faster.github.io/fold-wrapper"
+Issues = "https://github.com/dream-faster/fold-wrapper/issues"
+Source = "https://github.com/dream-faster/fold-wrapper"
 
 [project.optional-dependencies]
 quality = [
@@ -80,7 +80,7 @@ lightgbm = [
 
 [tool.hatch.envs.quality]
 dependencies = [
-  "fold-models[quality]",
+  "fold-wrapper[quality]",
 ]
 detached = true
 
@@ -97,7 +97,7 @@ format = [
 ]
 
 [tool.hatch.envs.test]
-dependencies = ["fold-models[tests]"]
+dependencies = ["fold-wrapper[tests]"]
 
 [tool.hatch.envs.test.scripts]
 run = "pytest tests/ --durations 0 -s"
